+++ CHANGELOG.md
@@ -8,9 +8,14 @@
   - **Multiple blockchecks.** It attempts to find the most suitable domain name for blockcheck.
   - **Atomic systems support.** Respect for atomic systems.
   - **Runit init support.** Respect for Runit init.
+  - **S6 init support.** Respect for S6 init.
   - **OpenRC init support.** Respect for OpenRC init.
   - **OpenBSD init support.** Respect for OpenBSD init.
+  - **FreeBSD init support.** Respect for FreeBSD init.
+  - **pfSense init support.** Respect for pfSense init.
   - **SysVinit init support.** Respect for SysVinit init.
+  - **Entware and Optware init support.** Respect for Entware and Optware init.
+  - **Launchd init support.** Respect for Launchd init.
   - **RPM-OSTREE package manager support.** Respect for RPM-OSTREE package manager.
   - **XBPS package manager support.** Respect for XBPS package manager.
   - **Emerge package manager support.** Respect for Emerge package manager.